% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btblv-package.R
\docType{package}
\name{btblv-package}
\alias{btblv-package}
\alias{btblv}
\title{The 'btblv' package.}
\description{
Package to fit time-dependent beta latent variable models
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
Pedro Menezes de Araujo
}
