% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btblv_fit.R
\name{btblv_fit}
\alias{btblv_fit}
\title{Fit time-dependent beta latent variable model}
\usage{
btblv_fit(
  btblv_data,
  K,
  precision = "single",
  iter,
  warmup,
  thin,
  chains,
  cores,
  seed,
  ...
)
}
\arguments{
\item{btblv_data}{object btblv_data generated by the function \code{create_btblv_data}}

\item{K}{interger with the latent dimension size}

\item{precision}{string}

\item{iter}{integer with the number of iterations}

\item{warmup}{integer with the warmup size}

\item{thin}{integer with the thinning size}

\item{chains}{integer with the number of chains}

\item{cores}{integer with the number of cores running the chains}

\item{seed}{integer with the random seed}

\item{...}{additional parameters for the \code{rstan::sampling} function}
}
\value{
object of the class btblv_fit
}
\description{
Fit time-dependent beta latent variable model
}
\examples{
##
}
