% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cWAIC.R
\name{calc_cWAIC}
\alias{calc_cWAIC}
\title{Compute WAIC based on the conditional likelihood}
\usage{
calc_cWAIC(btblv_posterior)
}
\arguments{
\item{btblv_posterior}{btblv_posterior btblv_posterior objected generated by \code{extract_posterior}.}
}
\value{
WAIC object from \code{loo::waic} package.
}
\description{
Compute WAIC based on the conditional likelihood
}
\examples{
#
}
