% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Compute converge checks for posterior samples from the btblv model}
\usage{
check_convergence(btblv_posterior)
}
\arguments{
\item{btblv_posterior}{\code{btblv_posterior} object generated by \code{extract_posterior}.}
}
\value{
data.frame with the the converge metrics
}
\description{
Compute converge checks for posterior samples from the btblv model
}
\examples{
#
}
