% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fit.R
\name{check_fit}
\alias{check_fit}
\title{Check model fit}
\usage{
check_fit(posterior_predict, posterior_summary)
}
\arguments{
\item{posterior_predict}{\code{posterior_predict} object generated by \code{posterior_predict}.}

\item{posterior_summary}{\code{posteior_summary} object generated by \code{posterior_summary}.}
}
\value{
list with the elements:
\itemize{
\item \code{global_metrics} - metrics for whole data.
\item \code{item_metrics} - metrics grouped by item.
\item \code{group_metrics} - metrics grouped by group.
\item \code{distance_metrics} - metrics for the distance between the observations.
}
}
\description{
Computes metrics for model fit like reconstruction RMSE, MAE and MAPE.
}
\examples{
#
}
