% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_WAIC.R
\name{compute_WAIC}
\alias{compute_WAIC}
\title{Compute WAIC based on the conditional likelihood}
\usage{
compute_WAIC(btblv_posterior)
}
\arguments{
\item{btblv_posterior}{btblv_posterior btblv_posterior objected generated by \code{extract_posterior}.}
}
\value{
WAIC object from \code{loo::waic} package.
}
\description{
Compute WAIC based on the conditional likelihood
}
\examples{
#
}
