% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btblv_data.R
\name{create_btblv_data}
\alias{create_btblv_data}
\title{Create object from the class btblv_data}
\usage{
create_btblv_data(
  df,
  resp_col_name,
  item_col_name,
  group_col_name,
  time_col_name
)
}
\arguments{
\item{df}{\code{data.frame} with at least \code{resp_column}, \code{item_column}, \code{group_column}, \code{time_column}.}

\item{resp_col_name}{string with the name of the column with the (0, 1) data.}

\item{item_col_name}{string with the name of the column with the items.}

\item{group_col_name}{string with the name of the columns with the groups.}

\item{time_col_name}{string with the name of the columns with the time.}
}
\value{
btblv_data object.
}
\description{
Create object with the data to be used in \code{btblv_fit}
}
\examples{
data("hmad_data")
lf = hmd_data$life_tables_5x5 |>
 dplyr::filter(year \%in\% seq(1990, 2020, 5))
data = create_btblv_data(df = lf,
                         resp_col_name = "mx",
                         item_col_name = "age",
                         group_col_name = "country",
                         time_col_name = "year")


}
