% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_last_dimension}
\alias{.add_last_dimension}
\title{Preserve the last shape in an array with 3 dimensions}
\usage{
.add_last_dimension(array)
}
\arguments{
\item{array}{array with two dimensions (n, J)}
}
\value{
array with dimension (n, J, 1)
}
\description{
Preserve the last shape in an array with 3 dimensions
}
\examples{
#
}
