% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.apply_increment_to_series}
\alias{.apply_increment_to_series}
\title{Generate AR1 series from increments of a group stacked increments}
\usage{
.apply_increment_to_series(E, group_index, sigma, phi)
}
\arguments{
\item{E}{matrix with the increments for each group stacked.}

\item{group_index}{numeric vector with the same size of the number of lines
of E. It indicates the group of the elements of E.}

\item{sigma}{numeric vector the sigma parameter for each group.}

\item{phi}{numeric vector the phi parameter for each group.}
}
\value{
numeric matrix with serie E.
}
\description{
Generate AR1 series from increments of a group stacked increments
}
\examples{
#
}
