% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.apply_rotation}
\alias{.apply_rotation}
\title{Apply the Procrustean transformation for each element of the posterior}
\usage{
.apply_rotation(post_sample, rotation_list)
}
\arguments{
\item{post_sample}{array with iteration, parameter posterior value, latent dimension.}

\item{rotation_list}{list with Procrustean transform generated by \code{.get_rotation}.}
}
\value{
array with the rotated posterior sample.
}
\description{
Apply the Procrustean transformation for each element of the posterior
}
\examples{
#
}
