% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compute_converge_metrics}
\alias{.compute_converge_metrics}
\title{Compute converge checks for a posterior sample}
\usage{
.compute_converge_metrics(array)
}
\arguments{
\item{array}{array with dimension (iters, chains, ...).}
}
\value{
data.frame with the metrics and id for first dimension, k for the second.
}
\description{
Computes the Rhat, ESS and ESS bulk for a posterior sample.
}
\examples{
#
}
