% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compute_MAE}
\alias{.compute_MAE}
\title{Mean absolute error}
\usage{
.compute_MAE(true, pred)
}
\arguments{
\item{true}{numeric vector}

\item{pred}{numeric vector}
}
\value{
numeric vector
}
\description{
Mean absolute error
}
\examples{
# .compute_RMSE(c(1, 2, 3), c(1, 2, 5))
}
