% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compute_MAPE}
\alias{.compute_MAPE}
\title{Mean absolite percentange error}
\usage{
.compute_MAPE(true, pred)
}
\arguments{
\item{true}{numeric vector}

\item{pred}{numeric vector}
}
\value{
numeric vector
}
\description{
Mean absolite percentange error
}
\examples{
# .compute_MAPE(c(1, 2, 3), c(1, 2, 5))

}
