% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compute_posterior_mean}
\alias{.compute_posterior_mean}
\title{Compute posterior mean from posterior sample rstan array}
\usage{
.compute_posterior_mean(array)
}
\arguments{
\item{array}{array with posterior sample.}
}
\value{
posterior mean.
}
\description{
Compute posterior mean from posterior sample rstan array
}
\examples{
#
}
