% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compute_RMSE}
\alias{.compute_RMSE}
\title{Root mean squared error}
\usage{
.compute_RMSE(true, pred)
}
\arguments{
\item{true}{numeric vector}

\item{pred}{numeric vector}
}
\value{
numeric vector
}
\description{
Root mean squared error
}
\examples{
# .compute_RMSE(c(1, 2, 3), c(1, 2, 5))
}
