% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_lower_tri}
\alias{.get_lower_tri}
\title{Extract the lower triangular value of a matrix}
\usage{
.get_lower_tri(matrix)
}
\arguments{
\item{matrix}{matrix}
}
\value{
vector with the lower triangular values
}
\description{
Extract the lower triangular value of a matrix
}
\examples{
# matrix(rnorm(100), 10, 10) |> .get_lower_tri()

}
