% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_summary_df}
\alias{.get_summary_df}
\title{Title}
\usage{
.get_summary_df(post_array, cred_mass = 0.95)
}
\arguments{
\item{post_array}{array with posterior generated by rstan}

\item{cred_mass}{credible interval mass default 0.95}
}
\value{
data.frame with posterior summaries
}
\description{
Title
}
\examples{
#
}
