% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.increment_to_series}
\alias{.increment_to_series}
\title{Generate AR1 series from the increments and parameters}
\usage{
.increment_to_series(E, sigma, phi)
}
\arguments{
\item{E}{numeric matrix with random increments.}

\item{sigma}{numeric value with sigma parameter.}

\item{phi}{numeric value with phi parameter.}
}
\value{
numeric matrix with AR1 series.
}
\description{
Generate AR1 series from the increments and parameters
}
\examples{
#
}
