% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.inv_logit}
\alias{.inv_logit}
\title{Inverse logit function}
\usage{
.inv_logit(x)
}
\arguments{
\item{x}{numeric real value}
}
\value{
invese of the logit
}
\description{
Inverse logit function
}
\examples{
# .inv_logit(10)
}
