% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.reshape_chain_posterior}
\alias{.reshape_chain_posterior}
\title{Reshape non-permuted posterior to look similar to the shape of the permuted chains}
\usage{
.reshape_chain_posterior(chain_array, param_name, param_dims)
}
\arguments{
\item{chain_array}{array of non-permuted chain generated by \code{rstan::extrac(stan_fit, permuted = F)}.}

\item{param_name}{character with the name of the parameter you want to reshape.}

\item{param_dims}{numeric with the dimension of the parameter you want to reshape.}
}
\value{
list where each element is the posterior sample.
}
\description{
Reshape non-permuted posterior to look similar to the shape of the permuted chains
}
\examples{
# .reshape_chain_posterior()
}
