% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.RMSE}
\alias{.RMSE}
\title{Root mean squared error}
\usage{
.RMSE(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
numeric vector
}
\description{
Root mean squared error
}
\examples{
RMSE(c(1, 2, 3), c(1, 2, 5))
}
