% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.simulate_normal_matrix}
\alias{.simulate_normal_matrix}
\title{Simulate matrix from normal distribution}
\usage{
.simulate_normal_matrix(ncol, nrow, mu, sigma)
}
\arguments{
\item{ncol}{number of cols}

\item{nrow}{number of rows}

\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of the normal distribution}
}
\value{
random matrix
}
\description{
Simulate matrix from normal distribution
}
\examples{
# .simulate_normal_matrix(ncol=100, nrow=2, mu=10, sigma=3)
}
