% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_fit}
\alias{example_fit}
\title{List with some fits for examples}
\format{
\subsection{\code{example_fit}}{

A list with two data.frame with mortality data from the Human mortality Database
\describe{
\item{country_name}{Country name}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.mortality.org/}
}
\usage{
example_fit
}
\description{
A list with btblv_fit objects
}
\keyword{datasets}
