% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cond_log_like.R
\name{extract_cond_log_like}
\alias{extract_cond_log_like}
\title{Extract the pointwise condotional log-likelihood for the btblv model}
\usage{
extract_cond_log_like(btblv_posterior)
}
\arguments{
\item{btblv_posterior}{btblv_posterior object.}
}
\value{
matrix with the pointwise log-likelihood for each iteration of the posterior.
}
\description{
Extract the pointwise condotional log-likelihood for the btblv model
}
\examples{
#
}
