% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior.R
\name{extract_posterior}
\alias{extract_posterior}
\title{Extract sample posterior from btblv_fit object}
\usage{
extract_posterior(btblv_fit, add_btblv_fit = FALSE, alpha_reference = "PCA")
}
\arguments{
\item{btblv_fit}{btblv_fit object generated by \code{btblv_fit}.}

\item{add_btblv_fit}{logical value indicating if the object should contain the btblv_fit,
default is TRUE.}

\item{alpha_reference}{alpha reference matrix for the procrustean matching procedure.
It is "PCA" for the PCA eigen values applied to the logit(y),
"post mode" for the posterior mode, and a numeric matrix for a custom user provided.}
}
\value{
object of the class \code{btblv_posterior}.
}
\description{
Extract sample posterior from btblv_fit object
}
\examples{
#
}
