% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_btblv.R
\name{fit_btblv}
\alias{fit_btblv}
\title{Fit the time-dependent beta latent variable model with HMC}
\usage{
fit_btblv(
  btblv_data,
  K,
  precision = "single",
  iter,
  warmup,
  thin,
  chains,
  cores,
  seed,
  ...
)
}
\arguments{
\item{btblv_data}{\code{btblv_data} object generated by the function \code{create_btblv_data}.}

\item{K}{integer with the latent dimension size.}

\item{precision}{string with the type of precision. It is "single" if all
items have the same precision or "specific" if each item has its own precision
parameter. Default is "single".}

\item{iter}{integer with the number of iterations}

\item{warmup}{integer with the warm-up size. See \code{rstan::sampling}.}

\item{thin}{integer with the thinning size. See \code{rstan::sampling}.}

\item{chains}{integer with the number of chains. See \code{rstan::sampling}.}

\item{cores}{integer with the number of cores running the chains. See \code{rstan::sampling}.}

\item{seed}{integer with the random seed.}

\item{...}{additional parameters for the \code{rstan::sampling} function.}
}
\value{
object of the class \code{btblv_fit}.
}
\description{
Fit the time-dependent beta latent variable model with HMC
}
\examples{
##
}
