% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmd_data}
\alias{hmd_data}
\title{List with data mortality rates from the HMD}
\format{
\subsection{\code{hmd_data}}{

A list with two data.frame with mortality data from the Human mortality Database
\describe{
\item{country_name}{Country name}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.mortality.org/}
}
\usage{
hmd_data
}
\description{
A list with two data.frames...
}
\keyword{datasets}
