% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imifa_to_blv.R
\name{imifa_to_blv}
\alias{imifa_to_blv}
\title{Transform result \code{Result_IMIFA} object to a object similar to \code{btblv_posterior} object.}
\usage{
imifa_to_blv(btblv_data, imifa_result)
}
\arguments{
\item{btblv_data}{data used to fit the IMIFA model applied to \code{btblv::create_btblv_data}.}

\item{imifa_result}{\code{Result_IMIFA} object generated by \code{IMIFA::get_IMIFA_results}.}
}
\value{
object \code{fa_imifa_posterior} similar to \code{btblv_posterior}. This object
can inherits the methods \code{btblv::posterior_summary} and \code{btblv::posterior_predict}.
}
\description{
Transform result \code{Result_IMIFA} object to a object similar to \code{btblv_posterior} object.
}
\examples{
#
}
