% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse of the logit function}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{numeric value.}
}
\value{
numeric value inverse of the logit.
}
\description{
Inverse of the logit function
}
\examples{
inv_logit(10)
}
