% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lf}
\alias{lf}
\title{Mortality rates from the Human Mortality Database (HMD)}
\format{
data.frame \code{lf}
A data frame with 23088 rows and 8 columns:
\describe{
\item{year}{Country name}
\item{age}{age group x for [x, x+5)}
\item{mx}{central mortality rate}
\item{qx}{probability of death}
\item{openinterval}{bool indicating with the age group interval is +inf open}
\item{country_code}{country code}
\item{country}{country name}
\item{date_extract}{date of the extraction from HMD}
}
}
\source{
\url{https://www.mortality.org/}
}
\usage{
lf
}
\description{
Data with age-specific mortality rates and probability of death for
different countries over time from the human mortality database
}
\keyword{datasets}
