% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrmatrix.R
\name{plot_corrmatrix}
\alias{plot_corrmatrix}
\title{Plot correlation matrix}
\usage{
plot_corrmatrix(btblv_data, transform = btblv::logit)
}
\arguments{
\item{btblv_data}{\code{btblv_data} object create with \code{btblv::create_btblv_data}}

\item{transform}{function to transform the data. Default is \code{btblv::logit}.
If you do not want transformations use \code{NULL}.}
}
\value{
\code{ggplot2} plot with the correlation matrix.
}
\description{
Plot correlation matrix
}
\examples{
example_fit$single_K2$btblv_data |> plot_corrmatrix()

}
