% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_effects.R
\name{plot_latent_effects}
\alias{plot_latent_effects}
\title{PLot latent effects over time}
\usage{
plot_latent_effects(posterior_summary, highlight = NULL)
}
\arguments{
\item{posterior_summary}{\code{posterior_summary} object generated by \code{btblv::posterior_summary}.}

\item{highlight}{character vector with groups to be highlighted on the plot. Default is NULL.}
}
\value{
ggplot graph object
}
\description{
PLot latent effects over time
}
\examples{
# example_fit$specific_K1 |>
# extract_posterior() |>
# posterior_summary() |>
# plot_latent_effects()

}
