% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot Kendall's trend coefficient}
\usage{
plot_trend(btblv_data)
}
\arguments{
\item{btblv_data}{\code{btblv_data} object create with \code{btblv::create_btblv_data}}
}
\value{
\code{ggplot2} plot with the correlation matrix.
}
\description{
Plot Kendall's trend coefficient
}
\examples{
example_fit$single_K2$btblv_data |> plot_trend()

}
