% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summary.R
\name{posterior_summary}
\alias{posterior_summary}
\title{Compute posterior summary for the btblv_posterior object}
\usage{
posterior_summary(btblv_posterior)
}
\arguments{
\item{btblv_posterior}{object of the class btblv_posterior generated by \code{extract_posterior}}
}
\value{
posterior summary
}
\description{
Compute posterior summary for the btblv_posterior object
}
\examples{
#
}
