% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.btblv_data.R
\name{print.btblv_data}
\alias{print.btblv_data}
\title{Print method for btblv_data object}
\usage{
\method{print}{btblv_data}(btblv_data, ...)
}
\arguments{
\item{btblv_data}{object from the class btblv_data}

\item{...}{extra arguments}
}
\value{
print message
}
\description{
Displays information about the the btblv_data object
}
\examples{
data("lf")
data = create_btblv_data(lf, "mx", "age", "country", "year")
data

}
