% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Generate data from the btblv model based on the posterior mean of the parameters}
\usage{
simulate_data(btblv_posterior, replicates, seed)
}
\arguments{
\item{btblv_posterior}{object of the class \code{btblv_posterior} generated by \code{extract_posterior}.}

\item{replicates}{integer value with the number of replicates.}

\item{seed}{int seed for the random values.}
}
\value{
list
}
\description{
Generate data from the btblv model based on the posterior mean of the parameters
}
\examples{
#
}
