% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btblv-package.R
\docType{package}
\name{btblv-package}
\alias{btblv-package}
\alias{btblv}
\title{The 'btblv' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
