% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.simulate_uniform_matrix}
\alias{.simulate_uniform_matrix}
\title{Simulate matrix from a uniform distribution}
\usage{
.simulate_uniform_matrix(ncol, nrow, lower, upper)
}
\arguments{
\item{ncol}{number of cols.}

\item{nrow}{number of rows.}

\item{lower}{min.}

\item{upper}{max.}
}
\value{
numeric matrix.
}
\description{
Simulate matrix from a uniform distribution
}
\examples{
# .simulate_uniform_matrix(ncol=2, nrow=100, lower=-1, upper=1)
}
