% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior.R
\name{extract_posterior}
\alias{extract_posterior}
\title{Extract sample posterior from btblv_fit object}
\usage{
extract_posterior(
  btblv_fit,
  alpha_reference = "pca",
  apply_varimax = FALSE,
  add_btblv_fit = FALSE
)
}
\arguments{
\item{btblv_fit}{btblv_fit object generated by \code{btblv_fit}.}

\item{alpha_reference}{alpha reference matrix for the Procrustean matching procedure.
It is "pca" for the pca eigenvalues applied to the logit(y), and
"mode" for the posterior mode, and a numeric matrix for a custom user provided.
Default is "pca".}

\item{apply_varimax}{logical value. if \code{TRUE}, applies a varimax transform to the reference alpha.
Default is \code{FALSE}.}

\item{add_btblv_fit}{logical value indicating if the object should contain the \code{btblv_fit} object.
Default is TRUE.}
}
\value{
object of the class \code{btblv_posterior} with the following attributes:
\describe{
\item{post_sample_array}{List where each element is an arrray with the permuted posterior distribution for the parameters.}
\item{post_sample_chains}{List where each element is the an array with the posterior distribution by chain for each parameter.}
\item{rotations}{List with all \code{vegan::procrustes} results for each iteration of the posterior.}
\item{btblv_data}{\code{btblv_data} object use to fit the model.}
\item{precision}{type of precision of the fit.}
\item{is_varimax}{\code{TRUE} if varimax transformed was applied.}
\item{alpha_reference}{the reference matrix used for the procrustes rotation.}
}
}
\description{
Extract sample posterior from btblv_fit object
}
\examples{
#
}
