% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior.R
\name{extract_posterior}
\alias{extract_posterior}
\title{Extract sample posterior from btblv_fit object}
\usage{
extract_posterior(btblv_fit, add_btblv_fit = FALSE)
}
\arguments{
\item{btblv_fit}{btblv_fit object generated by \code{btblv_fit}.}

\item{add_btblv_fit}{logical value indicating if the object should contain the btblv_fit,
default is TRUE.}
}
\value{
object of the class \code{btblv_posterior}.
}
\description{
Extract sample posterior from btblv_fit object
}
\examples{
#
}
