% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior.R
\name{extract_posterior}
\alias{extract_posterior}
\title{Extract sample posterior from btblv_fit object}
\usage{
extract_posterior(
  btblv_fit,
  alpha_reference = "pca",
  apply_varimax = FALSE,
  add_btblv_fit = FALSE
)
}
\arguments{
\item{btblv_fit}{btblv_fit object generated by \code{btblv_fit}.}

\item{alpha_reference}{alpha reference matrix for the Procrustean matching procedure.
It is "pca" for the pca eigenvalues applied to the logit(y), and
"mode" for the posterior mode, and a numeric matrix for a custom user provided.
Default is "pca".}

\item{apply_varimax}{logical value. if \code{TRUE}, applies a varimax transform to the reference alpha.
Default is \code{FALSE}.}

\item{add_btblv_fit}{logical value indicating if the object should contain the \code{btblv_fit} object.
Default is TRUE.}
}
\value{
object of the class \code{btblv_posterior}.
}
\description{
Extract sample posterior from btblv_fit object
}
\examples{
#
}
