% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summary.R
\name{posterior_summary}
\alias{posterior_summary}
\title{Compute posterior summary for the \code{btblv_posterior} object}
\usage{
posterior_summary(btblv_posterior, cred_mass = 0.95, pars = NULL)
}
\arguments{
\item{btblv_posterior}{object of the class \code{btblv_posterior}
generated by \code{btblv::extract_posterior}.}

\item{cred_mass}{numeric value between 0 and 1.}
}
\value{
object of the class \code{btblv_posterior_summary} with the attributes:
}
\description{
Compute posterior summary for the \code{btblv_posterior} object
}
\examples{
#
}
