% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.btblv_data.R
\name{print.btblv_data}
\alias{print.btblv_data}
\title{Print method for \code{btblv_data} object}
\usage{
\method{print}{btblv_data}(x, ...)
}
\arguments{
\item{x}{\code{btblv_data} object generate by \code{btblv::btblv_data}.}

\item{...}{extra arguments.}
}
\value{
message.
}
\description{
Displays information about the the \code{btblv_data} object.
}
\examples{
#

}
